Allow meta in addons.ini file
This minor change will allow a user to add the following two url's for channels in their addons.ini file:

plugin://plugin.video.meta/tv/play_by_name
plugin://plugin.video.meta/movies/play_by_name

Allowing so means that a user could, theoretically, add a channel such as Syfy or a movie channel, and when clicking on an episode, the 'channel stream' will automatically pick it up in Meta (so long as the user is using the correct season/episode data format in their xmltv files).

A flaw is that it will only play one episode, then return to the guide, and you'd have to click on the following episode, meaning the user would lose that 'binge watching a channel' experience. We can look into ways of automatically queuing up the following episodes/movies in a playlist later, should you like this change.

Thanks,
Ray.