Better auto-swicth channel with ID input
This is a better way to change channels on ID input. If you have a fullscreen video playing, it will automatically play the channel you input. However, if you are in the EPG it will not automatically play, instead just highlighting the channel. This assumes that the end user switching when in fullscreen video definitely wants to switch, where as in the EPG screen is still just browsing and a switch isn't totally necessary.