AltChannel and Recording Update
Provide an alternate background for the channel column in the guide. If the file tvg-alt-channel-column.png exists it will be used.

Allow for notification/autoplay/autoplaywith of only new episodes of a TV show.

Move autoplaywith recordings into seperate folders for Movies and TV Shows to provide better library integration.